requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
bcrypt==4.3.0
flask-socketio==5.5.1
eventlet==0.33.3 